+++ src/app/page.tsx
@@ -2,11 +2,13 @@ import { Button } from '@/components/feature/Button';
 import { Container } from '@/components/ui/Container';
 import Link from 'next/link';
 import pkg from 'betting_app/package.json';
+import { Logo } from '@/components/Logo';
+import { GitHub, Instagram, Twitter } from '@mui/icons-material';
 
 export default async function Home() {
   return (
-    <main className='lg:px-default xs:px-4 py-8 flex flex-col flex-1 text-accent gap-16 overflow-y-scroll'>
-      <section className='flex flex-col gap-4'>
+    <main className='pt-8 flex flex-col flex-1 gap-16 overflow-y-scroll'>
+      <Section>
         <Heading>{pkg.name} - A poker inspired betting app.</Heading>
         <p>
           Welcome to {pkg.name} — where the fun never stops and the bets are always on! Whether it’s
@@ -15,8 +17,13 @@ export default async function Home() {
           placed using "Dice," our quirky virtual currency, and the best part? It’s totally free to
           play! So, get your friends together, throw down some fun challenges, and see who’s got the
           best instincts.
+          <br />
+          <br />
+          <i className='text-sm text-gray-500'>
+            A limited amount of users can register at this time, due to resource constraints.
+          </i>
         </p>
-        <div className='xs:w-full lg:w-[25%]'>
+        <div className='xs:w-full lg:w-[25%] flex-col'>
           <Link href='/register'>
             <Button
               fullWidth
@@ -26,91 +33,127 @@ export default async function Home() {
             </Button>
           </Link>
         </div>
-      </section>
+      </Section>
 
-      <section className='flex flex-col gap-4'>
+      <Section>
         <Heading>Features</Heading>
         <ul className='grid grid-cols-1 gap-2'>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>Virtual Currency</h1>
+            <ContainerBody>
+              <ContainerHeading>Virtual Currency</ContainerHeading>
               <p>
                 Place your bets with <strong>Dice</strong>—our quirky virtual currency that lets you
                 join in on any bet! There’s no limit to how many bets you can make, and don’t worry,
                 balances can even go negative. It’s all part of the fun!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>Expiry Dates</h1>
+            <ContainerBody>
+              <ContainerHeading>Expiry Dates</ContainerHeading>
               <p>
                 Some bets are a race against time! You can set an expiry date for your bets, so they
                 stay open until the clock runs out. No rush—let the tension build!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>Fixed bid to enter</h1>
+            <ContainerBody>
+              <ContainerHeading>Fixed bid to enter</ContainerHeading>
               <p>
                 Every bet has a fixed amount you’ll need to wager to get in the game. No surprises
                 here—just place your bet and you’re part of the action!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
         </ul>
-      </section>
+      </Section>
 
-      <section className='flex flex-col gap-4'>
+      <Section>
         <Heading>Planned Features</Heading>
         <ul className='grid grid-cols-1 gap-2'>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>Bid raising</h1>
+            <ContainerBody>
+              <ContainerHeading>Bid raising</ContainerHeading>
               <p>
                 Get ready to raise the stakes! Soon, you’ll be able to increase your bid after
                 someone else has placed theirs, just like in poker. Think you can outbid your
                 friends? Prove it!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>A jury for closing bets</h1>
+            <ContainerBody>
+              <ContainerHeading>A jury for closing bets</ContainerHeading>
               <p>
                 Right now, the creator of a bet has the final say in closing it. But we’re working
                 on a feature that’ll let trusted users vote on the outcome, making things more
                 exciting and fair.
               </p>
-            </div>
+            </ContainerBody>
           </Container>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>A reputation score</h1>
+            <ContainerBody>
+              <ContainerHeading>A reputation score</ContainerHeading>
               <p>
                 Let your track record speak for itself! We’re planning to introduce a reputation
                 score that reflects your betting skills and your contribution as a jury member or
                 creator. The more trusted you are, the higher your score!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
           <Container as='li'>
-            <div className='flex flex-col gap-2 text-accent'>
-              <h1 className='font-semibold'>Custom Currencies</h1>{' '}
+            <ContainerBody>
+              <ContainerHeading>Custom Currencies</ContainerHeading>{' '}
               <p>
                 Why stop at Dice? Soon, you’ll be able to create your own custom currencies for even
                 more fun! Want to bet with “Gems,” “Coins,” or something totally unique? The choice
                 is yours. Make the game feel truly yours!
               </p>
-            </div>
+            </ContainerBody>
           </Container>
         </ul>
-      </section>
+      </Section>
+      <footer className='flex w-full py-8 bg-accent lg:px-default xs:px-4 text-white flex-col'>
+        <div className='flex flex-col gap-2 pb-12'>
+          <Link
+            href='/tos'
+            target='_blank'>
+            <span className='text-white font-semibold'>Terms Of Service</span>
+          </Link>
+          <Link
+            target='_blank'
+            href='/privacy'>
+            <span className='text-white font-semibold'>Privacy Policy</span>
+          </Link>
+        </div>
+        <div className='flex justify-between items-center'>
+          <Logo />
+          <div className='flex gap-4'>
+            <Link
+              target='_blank'
+              href='https://github.com/Oletusarvo/wagerize'>
+              <GitHub sx={{ color: 'white' }} />
+            </Link>
+          </div>
+        </div>
+      </footer>
     </main>
   );
 }
 
 function Heading({ as = 'h1', children, ...props }) {
-  return <h1 className='text-2xl font-semibold'>{children}</h1>;
+  return <h1 className='text-2xl font-semibold text-accent'>{children}</h1>;
+}
+
+function ContainerHeading({ children }) {
+  return <h1 className='text-accent font-semibold'>{children}</h1>;
+}
+
+function ContainerBody({ children }) {
+  return <div className='flex flex-col gap-2'>{children}</div>;
+}
+
+function Section({ children }) {
+  return <section className='flex flex-col gap-4 lg:px-default xs:px-4'>{children}</section>;
 }