+++ src/features/users/register/types/RegisterError.ts
@@ -1,3 +1,4 @@
 export enum RegisterError {
   DUPLICATE_USER = 'duplicate_user',
+  USER_COUNT = 'user_count',
 }